// src/lib/server/graphqlServer/types/generated-resolvers.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type AccountSort = {
  accountGroup?: InputMaybe<SortDirection>;
  accountGroup2?: InputMaybe<SortDirection>;
  accountGroup3?: InputMaybe<SortDirection>;
  accountGroupCombined?: InputMaybe<SortDirection>;
  accountTitleCombined?: InputMaybe<SortDirection>;
  active?: InputMaybe<SortDirection>;
  allowUpdate?: InputMaybe<SortDirection>;
  deleted?: InputMaybe<SortDirection>;
  disabled?: InputMaybe<SortDirection>;
  endDate?: InputMaybe<SortDirection>;
  isCash?: InputMaybe<SortDirection>;
  isNetWorth?: InputMaybe<SortDirection>;
  startDate?: InputMaybe<SortDirection>;
  status?: InputMaybe<SortDirection>;
  title?: InputMaybe<SortDirection>;
};
// ...
export type SortDirection =
  | 'asc'
  | 'desc';